Suggestion to strike "3 days" in favor of "a date"
My thinking is, what if we settle on 2 days? Do we make a new ADR
amending this one? I'd rather that detail be left to convention and
group norms, maybe to be discussed in retrospectives, but not really in
scope for the decision to adopt lazy consensus.

This way, there's a date, and it's part of the proposal.